fix read-only bug
