+++ .travis.yml
@@ -13,7 +13,7 @@ matrix:
 env:
   global:
     - ES_VER=1.4.2
-    - ES_MAPPER_ATTACHMENTS_VER=2.0.0.RC1
+    - ES_MAPPER_ATTACHMENTS_VER=2.4.1
     - ES_TRANSPORT_THRIFT_VER=2.0.0
     - ES_GEOCLUSTER_FACET_VER=0.0.10
     - ES_WAIT_ON_MAPPING_CHANGE=true
