+++ test/lib/Elastica/Test/Index/SettingsTest.php
@@ -10,7 +10,7 @@ use Elastica\Test\Base as BaseTest;
 
 class SettingsTest extends BaseTest
 {
-	
+
 	public function testGet()
     {
         $indexName = 'elasticatest';
@@ -197,12 +197,11 @@ class SettingsTest extends BaseTest
 
         $type = $index->getType('test');
         $type->addDocument($doc1);
-        $this->assertEquals('false', $index->getSettings()->get('blocks.read_only')); //ES returns a string for this setting
+        $this->assertFalse($index->getSettings()->getReadOnly());
 
         // Try to add doc to read only index
         $index->getSettings()->setReadOnly(true);
-
-        $this->assertEquals('true', $index->getSettings()->get('blocks.read_only'));
+        $this->assertTrue($index->getSettings()->getReadOnly());
 
         try {
             $type->addDocument($doc2);
@@ -210,7 +209,7 @@ class SettingsTest extends BaseTest
         } catch (ResponseException $e) {
             $message = $e->getMessage();
             $this->assertContains('ClusterBlockException', $message);
-            $this->assertContains('index read-only', $message);
+            $this->assertContains('index write', $message);
         }
 
         // Remove read only, add document
@@ -285,18 +284,5 @@ class SettingsTest extends BaseTest
 
         $settings->setBlocksMetadata(false); // Cannot delete index otherwise
         $index->delete();
-    }
-
-    protected function _waitForAllocation(Index $index)
-    {
-        do {
-            $settings = $index->getStatus()->get();
-            $allocated = true;
-            foreach ($settings['shards'] as $shard) {
-                if ($shard[0]['routing']['state'] != 'STARTED') {
-                    $allocated = false;
-                }
-            }
-        } while (!$allocated);
     }
 }
